Stats page to new format (#11)
* Stats page to new format

Along the way I realized that I rewrote a bunch of stuff, but might have lost sight about if things are better. I let you guys weigh in on if it's worth it. If not, we can roll this back and the diff before it. I /think/ it'll make it easier to build a player page and do more with the data going forward.

- Move stats.js to calculate stuff in getStaticProps. This essentially moves the logic of generateData.js into getStaticProps and various util methods.
- Make player info into an object so we can store other info in there (notably, [fb]id's). This will allow us to grab profile pics, and are better url safe handles for player profiles.
- "Fix" (I think) the asc/desc logic for sorting stats. The old arrows didn't match my understanding of asc/desc, but I could be wrong.
- Lots of lodash usage to avoid rewriting those functions. Lmk if this is bad. I think it makes the code a bit easier to read.

* Cleaning up includes