Add support for skins (and some bonuses)
Closes #15

This diff adds support for skins into the system. That means a marker in the spreadsheet to note that a player played skins ("(s)"), storing this data in round data, and rendering it out when show the results of a single round.

While updated the spreadsheet parsing, I added a couple features.
- Cell A1 can now be used for the name of the course.
- There's an optional row that can be added below the list of holes that will list the par values for those holes.
- The individual score objects now include the par value for the hole.
These changes are a big step towards #13. In fact, any 9 hole course should be supported now. We'd want to do the additional step of making the UI separate stats for different course before adding data from other places.